reorganize files into groups which reflect MVC architecture.
Move existing files into appropriate folders, and separate
utility extensions into their own files.
