+++ Misc/PathLogger.swift
